+++ aio_overpass/pt.py
@@ -103,7 +103,6 @@ class SingleRouteQuery(RouteQuery):
         super().__init__(input_code, **kwargs)
 
 
-@dataclass
 class RoutesWithinQuery(RouteQuery):
     """
     A query that produces ``Route`` objects for any route within the exterior of a polygon.
@@ -177,7 +176,7 @@ class Connection(Enum):
         return f"{type(self).__name__}.{self.name}"
 
 
-@dataclass
+@dataclass(slots=True)
 class Stop(Spatial):
     """
     A stop on a public transportation route.
@@ -347,7 +346,7 @@ class RouteScheme(Enum):
         return f"{type(self).__name__}.{self.name}"
 
 
-@dataclass
+@dataclass(slots=True)
 class Route(Spatial):
     """
     A public transportation service route, e.g. a bus line.